% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_area}
\alias{calc_area}
\title{Calculate Area
Ensure it's converted to km^2 then dropped}
\usage{
calc_area(.data)
}
\description{
Calculate Area
Ensure it's converted to km^2 then dropped
}
\keyword{internal}
