% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_n1_sum}
\alias{calc_n1_sum}
\title{Calculate conc sum from n1 to n
\deqn{\sum_{i=1}^{n1}\frac{t_i*a_i}{T_1}}}
\usage{
calc_n1_sum(.data, .X)
}
\description{
Calculate conc sum from n1 to n
\deqn{\sum_{i=1}^{n1}\frac{t_i*a_i}{T_1}}
}
\keyword{internal}
