% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calc_n2_sum}
\alias{calc_n2_sum}
\title{Calculate conc sum from n2 to n
\deqn{\sum_{i=n2}^{n}\frac{t_i*a_i}{T_2}}}
\usage{
calc_n2_sum(.data, .X)
}
\description{
Calculate conc sum from n2 to n
\deqn{\sum_{i=n2}^{n}\frac{t_i*a_i}{T_2}}
}
\keyword{internal}
