% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{de_county}
\alias{de_county}
\title{de_county}
\format{
An sf dataframe with 3 observations
}
\usage{
data("de_county")
}
\description{
This data contains 2010 Census data for each of the three counties in DE.
}
\examples{
data("de_county")
}
\concept{data}
