% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{de_tract}
\alias{de_tract}
\title{de_tract}
\format{
An sf dataframe with 218 observations
}
\usage{
data("de_tract")
}
\description{
This data contains 2010 Census data for each of the 218 tracts in DE.
}
\examples{
data("de_tract")
}
\concept{data}
