% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissim.R
\name{ds_dissim}
\alias{ds_dissim}
\title{Compute Dissimilarity Index}
\usage{
ds_dissim(.data, .cols, .name, .comp = FALSE)
}
\arguments{
\item{.data}{\link[tibble:tibble-package]{tibble}}

\item{.cols}{\href{https://tidyselect.r-lib.org/reference/language.html}{\code{tidy-select}}
Columns to compute the measure with. Must be at least 2 columns. If more than 2, treats
first column as first group and sum of other columns as second.}

\item{.name}{name for column with dissimilarity index. Leave missing to return a vector.}

\item{.comp}{Default is FALSE. FALSE returns the sum, TRUE returns the components.}
}
\value{
a \link[tibble:tibble-package]{tibble} or numeric vector if .name missing
}
\description{
Compute Dissimilarity Index
}
\examples{
data("de_county")
ds_dissim(de_county, c(pop_white, starts_with('pop_')))
ds_dissim(de_county, c(pop_white, starts_with('pop_')), .comp = TRUE)
ds_dissim(de_county, starts_with('pop_'), 'dissim')
}
\concept{seg}
