% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhi.R
\name{ds_hhi}
\alias{ds_hhi}
\title{Compute Herfindahl-Hirshman Index}
\usage{
ds_hhi(.data, .cols, .name)
}
\arguments{
\item{.data}{\link[tibble:tibble-package]{tibble}}

\item{.cols}{\href{https://tidyselect.r-lib.org/reference/language.html}{\code{tidy-select}}
Columns to compute the measure with.}

\item{.name}{name for column with HHI. Leave missing to return a vector.}
}
\value{
a \link[tibble:tibble-package]{tibble} or numeric vector if .name missing
}
\description{
Compute Herfindahl-Hirshman Index
}
\examples{
data("de_county")
ds_hhi(de_county, starts_with('pop_'))
ds_hhi(de_county, starts_with('pop_'), 'blau')
}
\concept{div}
